feat: auth error handler + separated logic
